Update README.md
Bright Ideas is a social blogging platform where users can log in, share their ideas, and interact with others by liking posts. The platform allows users to see the most liked ideas at the top of the feed and restricts users from deleting ideas posted by others. This project is built with Java and Spring Boot for the backend, JSP and Bootstrap for the frontend, and MySQL as the database.